requirements.txt
Babel == 2.8.0
certifi == 2020.6.20
cycler == 0.10.0
kiwisolver == 1.2.0
matplotlib == 3.3.1
multitimer == 0.2
numpy == 1.19.1
paho-mqtt == 1.5.0
pandas == 1.1.1
Pillow == 7.2.0
pip == 19.0.3
psycopg2 == 2.8.5
pyparsing == 2.4.7
python-dateutil == 2.8.1
pytz == 2020.1
scipy == 1.5.3
setuptools == 40.8.0
six == 1.15.0
tkcalendar == 1.6.1
xlrd == 1.2.0
